[move]GenreCategoryコンポーネントのディレクトリ構成を変更
