+++ resources/js/app.ts
@@ -3,7 +3,6 @@ import '../css/app.css'
 
 import { createApp } from "vue/dist/vue.esm-bundler";
 import CreatePlaylist from './feature/CreatePlaylist/CreatePlaylist.vue';
-import GenreCategory from './feature/GenreCategory/GenreCategory.vue';
 import { plugin, defaultConfig } from '@formkit/vue'
 import { createPinia } from "pinia";
 
@@ -12,7 +11,6 @@ const pinia = createPinia();
 const createPlaylistApp = createApp({
     components: {
         'create-playlist': CreatePlaylist,
-        'genre-category': GenreCategory,
     }
 });
 createPlaylistApp.use(plugin, defaultConfig);
